✨: Create new class to handle mongodb calls
