+++ libs/api/pdf-manager/repository/data-access/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/api-pdf-manager-repository-data-access.module';
+export * from './lib/pdf-manager-repository-data-access.module';
