+++ libs/api/pdf-manager/repository/data-access/src/lib/mongodb-access.spec.ts
@@ -1,0 +1,7 @@
+import { MongoDBAccess } from './mongodb-access';
+
+describe('MongoAccess', () => {
+  it('should be defined', () => {
+    expect(new MongoDBAccess()).toBeDefined();
+  });
+});
