+++ libs/api/pdf-manager/repository/data-access/src/lib/pdf-manager-repository-data-access.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
+import { MongoDBAccess } from './mongodb-access';
 
 @Module({
   controllers: [],
-  providers: [],
-  exports: [],
+  providers: [MongoDBAccess],
+  exports: [MongoDBAccess],
 })
 export class ApiPdfManagerRepositoryDataAccessModule {}
